refactoring - extract getLinksForOneDependencyFile() function to a new class DependencyFileLinkGenerator
